fix: use double for temparature

+++ main.py
@@ -17,7 +17,7 @@ parser.add_argument("--model", type=str, help="path or size of model")
 parser.add_argument("input_or_url", type=str, help="the file to transcribe")
 parser.add_argument("--pools", type=int, default=2, help="path to thumbnail")
 parser.add_argument(
-    "--temperature", type=int, default=0.5, help="convert wav file to opus"
+    "--temperature", type=double, default=0.5, help="convert wav file to opus"
 )
 parser.add_argument(
     "--wav_to_opus", type=int, default=True, help="convert wav file to opus"
